[Refactor] move decorator `IBrowserService` into common files to avoid potential file-level recursive dependency
